pyproject.toml
[project]
name = "kaggle-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "catboost>=1.2.7",
  "hydra-core>=1.3.2",
  "ipykernel>=6.29.5",
  "kaggle>=1.6.9",
  "lightgbm>=4.5",
  "lightning>=2.3.3",
  "mypy>=1.11.2",
  "pandas>=2.2.2",
  "polars>=1.7.1",
  "pre-commit>=3.8",
  "pyarrow>=17",
  "pydantic>=2.7.4",
  "pytest>=8.3.3",
  "rich>=13.8.1",
  "ruff>=0.6.5",
  "scikit-learn>=1.5.2",
  "seaborn>=0.13.2",
  "torch==2.0.1+cu118",
  "torch-geometric>=2.6",
  "transformers>=4.44.2",
  "wandb>=0.18.1",
  "xgboost>=2.1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["exp", "src"]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu118"]
